URL sem parâmetros, rotas login e controller padronizadas, login front padronizado.
