+++ client/src/1app/index.tsx
@@ -1,11 +1,10 @@
 import axios from 'axios'
 import ReactDOM from 'react-dom/client'
 import { realworld, handleGenericError } from '~6shared/api'
-import { sessionLib, useSessionStore } from '~6shared/session'
+import { useSessionStore } from '~6shared/session'
+import { GoogleLogin } from './auth'
 import { Provider } from './providers'
 import './main.css'
-// import { useLoginMutation } from '~4features/session/login/login.mutation'
-// import { BsWindowSidebar } from 'react-icons/bs'
 
 window.addEventListener('error', (event) => {
   if (axios.isAxiosError(event.error)) {
@@ -40,10 +39,4 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <Provider />,
 )
 
-const url_params = JSON.parse(decodeURIComponent(window.location.search.substring(1)).replace(/=([^&]+)\&/g, "=\"$1\",\n\r").replace(/(^|\r)([^=]+)=/g, "\"$2\":").replace(/^/, '{').replace(/$/, '}')) //eslint-disable-line
-
-const { setSession } = useSessionStore.getState();
-
-const session = sessionLib.transformUserDtoToSession(url_params);
-
-setSession(session);
+GoogleLogin();
