+++ client/src/4features/session/index.ts
@@ -1,4 +1,2 @@
-export * from './login'
 export * from './logout'
-export * from './register'
 export * from './update-session'
