+++ server/.env.example
@@ -64,3 +64,5 @@ SAIL_SKIP_CHECKS=true
 GOOGLE_CLIENT_ID=
 GOOGLE_CLIENT_SECRET=
 GOOGLE_REDIRECT_URL=http://127.0.0.1:${APP_PORT}/api/auth/callback
+
+FRONT_END_URL=http://127.0.0.1:5173
