+++ server/config/frontend.php
@@ -1,0 +1,2 @@
+<?php
+return ['url' => env("FRONT_END_URL"),];
